requirements.txt
pdfplumber>=0.10.0
openpyxl>=3.1.0
PyYAML>=6.0
